+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.2.9.0 -- 2022-??-??
 
 * Add `toShortText` and `toShortTextU` to `Data.Byte.Text.Ascii`.
+* Add `fromShortText` to `Data.Bytes.Text.Utf8`.
 
 ## 0.2.8.0 -- 2022-11-21